+++ extensions/google-chrome-profiles/CHANGELOG.md
@@ -1,13 +1,21 @@
 # Google Chrome Profiles Changelog
 
+## [Feature] - 2026-01-03
+
+- Bring the profile window to front instead of adding a new tab
+- Simplify deeplink integration
+- Improve bookmark favicon display
+
 ## [Refactor] - 2025-09-03
 
 - Refactor the extension to use the new `open-profile` and `open-profile-url` commands.
 
 ## [Quicklinks] - 2024-07-29
+
 - Add support for quicklinks to open a chosen profile in a specific url.
 
 ## [Quicklinks] - 2023-09-22
+
 - Support quicklinks to open a chosen profile.
 
 ## [Update] - 2023-02-13